Introduce AuthorizationFragment
