+++ common/src/main/java/com/microsoft/identity/common/internal/request/AcquireTokenOperationParameters.java
@@ -25,6 +25,8 @@ package com.microsoft.identity.common.internal.request;
 import android.app.Activity;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+
 import android.util.Pair;
 
 import com.google.gson.annotations.Expose;
@@ -38,6 +40,7 @@ import java.util.List;
 public class AcquireTokenOperationParameters extends OperationParameters {
 
     private transient Activity mActivity;
+    private transient Fragment mFragment;
     private String mLoginHint;
     @Expose()
     private List<Pair<String, String>> mExtraQueryStringParameters;
@@ -63,8 +66,16 @@ public class AcquireTokenOperationParameters extends OperationParameters {
         return mActivity;
     }
 
-    public void setActivity(@NonNull final Activity mActivity) {
-        this.mActivity = mActivity;
+    public void setActivity(@NonNull final Activity activity) {
+        this.mActivity = activity;
+    }
+
+    public Fragment getFragment() {
+        return mFragment;
+    }
+
+    public void setFragment(@NonNull final Fragment fragment) {
+        this.mFragment = fragment;
     }
 
     public List<Pair<String, String>> getExtraQueryStringParameters() {
