+++ common/src/main/java/com/microsoft/identity/common/internal/ui/AuthorizationStrategyFactory.java
@@ -23,6 +23,7 @@
 package com.microsoft.identity.common.internal.ui;
 
 import android.content.Context;
+
 import androidx.annotation.NonNull;
 
 import com.microsoft.identity.common.exception.ClientException;
@@ -51,13 +52,17 @@ public class AuthorizationStrategyFactory<GenericAuthorizationStrategy extends A
         //Valid if available browser installed. Will fallback to embedded webView if no browser available.
         final AuthorizationAgent validatedAuthorizationAgent = validAuthorizationAgent(
                 parameters.getAuthorizationAgent(),
-                parameters.getActivity().getApplicationContext()
+                parameters.getAppContext()
         );
         boolean isBrokerRequest = (parameters instanceof BrokerAcquireTokenOperationParameters);
 
         if (validatedAuthorizationAgent == AuthorizationAgent.WEBVIEW) {
             Logger.info(TAG, "Use webView for authorization.");
-            return (GenericAuthorizationStrategy) (new EmbeddedWebViewAuthorizationStrategy(parameters.getActivity()));
+            return (GenericAuthorizationStrategy) (
+                    new EmbeddedWebViewAuthorizationStrategy(
+                            parameters.getAppContext(),
+                            parameters.getActivity(),
+                            parameters.getFragment()));
         } else if (validatedAuthorizationAgent == AuthorizationAgent.DEFAULT) {
             // When the authorization agent is set to DEFAULT,
             // Use device browser auth flow as default.
@@ -67,12 +72,18 @@ public class AuthorizationStrategyFactory<GenericAuthorizationStrategy extends A
             } catch (final ClientException exception) {
                 Logger.info(TAG, "No supported browser available found. Fallback to the webView authorization agent.");
                 if (exception.getErrorCode().equalsIgnoreCase(ErrorStrings.NO_AVAILABLE_BROWSER_FOUND)) {
-                    return (GenericAuthorizationStrategy) (new EmbeddedWebViewAuthorizationStrategy(parameters.getActivity()));
+                    return (GenericAuthorizationStrategy) (
+                            new EmbeddedWebViewAuthorizationStrategy(
+                                    parameters.getAppContext(),
+                                    parameters.getActivity(),
+                                    parameters.getFragment()));
                 }
             }
             Logger.info(TAG, "Use browser for authorization.");
             final BrowserAuthorizationStrategy browserAuthorizationStrategy = new BrowserAuthorizationStrategy(
+                    parameters.getAppContext(),
                     parameters.getActivity(),
+                    parameters.getFragment(),
                     isBrokerRequest
             );
             browserAuthorizationStrategy.setBrowserSafeList(parameters.getBrowserSafeList());
@@ -80,7 +91,9 @@ public class AuthorizationStrategyFactory<GenericAuthorizationStrategy extends A
         } else {
             Logger.info(TAG, "Use browser for authorization.");
             final BrowserAuthorizationStrategy browserAuthorizationStrategy = new BrowserAuthorizationStrategy(
+                    parameters.getAppContext(),
                     parameters.getActivity(),
+                    parameters.getFragment(),
                     isBrokerRequest
             );
             browserAuthorizationStrategy.setBrowserSafeList(parameters.getBrowserSafeList());
